set double tensor

+++ irff/set_matrix_tensor.py
@@ -36,13 +36,13 @@ def set_matrix(m_,spec,bonds,mfopt,beopt,bdopt,messages,
                 m[pref+'bi_'+bd] = m[pref+'bi']
             elif pref+'wi_'+bd in m_ and reuse_m:                  # input layer
                 if bd in bdopt:
-                    m[pref+'wi_'+bd] = nn.Parameter(torch.tensor(m_[pref+'wi_'+bd]),
-                                                    requires_grad=True)
-                    m[pref+'bi_'+bd] = nn.Parameter(torch.tensor(m_[pref+'bi_'+bd]),
-                                                    requires_grad=True)
+                    m[pref+'wi_'+bd] = nn.Parameter(torch.tensor(m_[pref+'wi_'+bd],
+                                             dtype=torch.double),requires_grad=True)
+                    m[pref+'bi_'+bd] = nn.Parameter(torch.tensor(m_[pref+'bi_'+bd],
+                                                 dtype=torch.double),requires_grad=True)
                 else:
-                    m[pref+'wi_'+bd] = torch.tensor(m_[pref+'wi_'+bd])
-                    m[pref+'bi_'+bd] = torch.tensor(m_[pref+'bi_'+bd])
+                    m[pref+'wi_'+bd] = torch.tensor(m_[pref+'wi_'+bd],dtype=torch.double)
+                    m[pref+'bi_'+bd] = torch.tensor(m_[pref+'bi_'+bd],dtype=torch.double)
             else:
                 m[pref+'wi_'+bd] = nn.Parameter(torch.randn(nin,layer[0]),
                                                 requires_grad=True)  
@@ -58,14 +58,14 @@ def set_matrix(m_,spec,bonds,mfopt,beopt,bdopt,messages,
             elif pref+'w_'+bd in m_ and reuse_m:     
                 if bd in bdopt:                            
                     for i in range(layer[1]):   
-                        m[pref+'w_'+bd].append(nn.Parameter(torch.tensor(m_[pref+'w_'+bd][i]),
-                                               requires_grad=True) ) 
-                        m[pref+'b_'+bd].append(nn.Parameter(torch.tensor(m_[pref+'b_'+bd][i]),
-                                               requires_grad=True) ) 
+                        m[pref+'w_'+bd].append(nn.Parameter(torch.tensor(m_[pref+'w_'+bd][i],
+                                                     dtype=torch.double),requires_grad=True) ) 
+                        m[pref+'b_'+bd].append(nn.Parameter(torch.tensor(m_[pref+'b_'+bd][i],
+                                                    dtype=torch.double),requires_grad=True) ) 
                 else:
                     for i in range(layer[1]):   
-                        m[pref+'w_'+bd].append(torch.tensor(m_[pref+'w_'+bd][i]) )
-                        m[pref+'b_'+bd].append(torch.tensor(m_[pref+'b_'+bd][i])) 
+                        m[pref+'w_'+bd].append(torch.tensor(m_[pref+'w_'+bd][i],dtype=torch.double) )
+                        m[pref+'b_'+bd].append(torch.tensor(m_[pref+'b_'+bd][i],dtype=torch.double)) 
             else:
                 for i in range(layer[1]):   
                     m[pref+'w_'+bd].append(nn.Parameter(torch.randn(layer[0],layer[0]), 
@@ -78,13 +78,13 @@ def set_matrix(m_,spec,bonds,mfopt,beopt,bdopt,messages,
                 m[pref+'bo_'+bd] = m[pref+'bo']
             elif pref+'wo_'+bd in m_ and reuse_m:          # output layer
                 if bd in bdopt:       
-                    m[pref+'wo_'+bd] = nn.Parameter(torch.tensor(m_[pref+'wo_'+bd]),
-                                                    requires_grad=True) 
-                    m[pref+'bo_'+bd] = nn.Parameter(torch.tensor(m_[pref+'bo_'+bd]),
-                                                    requires_grad=True) 
+                    m[pref+'wo_'+bd] = nn.Parameter(torch.tensor(m_[pref+'wo_'+bd],
+                                             dtype=torch.double),requires_grad=True) 
+                    m[pref+'bo_'+bd] = nn.Parameter(torch.tensor(m_[pref+'bo_'+bd],
+                                             dtype=torch.double),requires_grad=True) 
                 else:
-                    m[pref+'wo_'+bd] = torch.tensor(m_[pref+'wo_'+bd])
-                    m[pref+'bo_'+bd] = torch.tensor(m_[pref+'bo_'+bd])
+                    m[pref+'wo_'+bd] = torch.tensor(m_[pref+'wo_'+bd],dtype=torch.double)
+                    m[pref+'bo_'+bd] = torch.tensor(m_[pref+'bo_'+bd],dtype=torch.double)
             else:
                 m[pref+'wo_'+bd] = nn.Parameter(torch.randn([layer[0],nout],stddev=0.2), 
                                                 requires_grad=True)    
@@ -109,27 +109,27 @@ def set_matrix(m_,spec,bonds,mfopt,beopt,bdopt,messages,
                 m[pref+'b_'+sp]  = m[pref+'b'] 
             elif pref+'wi_'+sp in m_ and reuse_m:
                 if sp in mfopt:
-                    m[pref+'wi_'+sp] = nn.Parameter(torch.tensor(m_[pref+'wi_'+sp]),
-                                                    requires_grad=True)
-                    m[pref+'bi_'+sp] = nn.Parameter(torch.tensor(m_[pref+'bi_'+sp]),
-                                                    requires_grad=True)
-                    m[pref+'wo_'+sp] = nn.Parameter(torch.tensor(m_[pref+'wo_'+sp]),
-                                                    requires_grad=True)
-                    m[pref+'bo_'+sp] = nn.Parameter(torch.tensor(m_[pref+'bo_'+sp]),
-                                                    requires_grad=True)
+                    m[pref+'wi_'+sp] = nn.Parameter(torch.tensor(m_[pref+'wi_'+sp],
+                                            dtype=torch.double),requires_grad=True)
+                    m[pref+'bi_'+sp] = nn.Parameter(torch.tensor(m_[pref+'bi_'+sp],
+                                            dtype=torch.double), requires_grad=True)
+                    m[pref+'wo_'+sp] = nn.Parameter(torch.tensor(m_[pref+'wo_'+sp],
+                                            dtype=torch.double),requires_grad=True)
+                    m[pref+'bo_'+sp] = nn.Parameter(torch.tensor(m_[pref+'bo_'+sp],
+                                            dtype=torch.double),requires_grad=True)
                     for i in range(layer[1]):   
-                        m[pref+'w_'+sp].append(nn.Parameter(torch.tensor(m_[pref+'w_'+sp][i]),
-                                               requires_grad=True)) 
-                        m[pref+'b_'+sp].append(nn.Parameter(torch.tensor(m_[pref+'b_'+sp][i]),
-                                               requires_grad=True)) 
+                        m[pref+'w_'+sp].append(nn.Parameter(torch.tensor(m_[pref+'w_'+sp][i],
+                                            dtype=torch.double),requires_grad=True)) 
+                        m[pref+'b_'+sp].append(nn.Parameter(torch.tensor(m_[pref+'b_'+sp][i],
+                                            dtype=torch.double),requires_grad=True)) 
                 else:
-                    m[pref+'wi_'+sp] = torch.tensor(m_[pref+'wi_'+sp])
-                    m[pref+'bi_'+sp] = torch.tensor(m_[pref+'bi_'+sp])
-                    m[pref+'wo_'+sp] = torch.tensor(m_[pref+'wo_'+sp])
-                    m[pref+'bo_'+sp] = torch.tensor(m_[pref+'bo_'+sp])
+                    m[pref+'wi_'+sp] = torch.tensor(m_[pref+'wi_'+sp],dtype=torch.double)
+                    m[pref+'bi_'+sp] = torch.tensor(m_[pref+'bi_'+sp],dtype=torch.double)
+                    m[pref+'wo_'+sp] = torch.tensor(m_[pref+'wo_'+sp],dtype=torch.double)
+                    m[pref+'bo_'+sp] = torch.tensor(m_[pref+'bo_'+sp],dtype=torch.double)
                     for i in range(layer[1]):   
-                        m[pref+'w_'+sp].append(torch.tensor(m_[pref+'w_'+sp][i] ) ) 
-                        m[pref+'b_'+sp].append(torch.tensor(m_[pref+'b_'+sp][i] ) ) 
+                        m[pref+'w_'+sp].append(torch.tensor(m_[pref+'w_'+sp][i],dtype=torch.double ) ) 
+                        m[pref+'b_'+sp].append(torch.tensor(m_[pref+'b_'+sp][i],dtype=torch.double ) ) 
             else:
                 m[pref+'wi_'+sp] = nn.Parameter(torch.randn(nin,layer[0]))   
                 m[pref+'bi_'+sp] = nn.Parameter(torch.randn(layer[0]))  
@@ -156,25 +156,29 @@ def set_matrix(m_,spec,bonds,mfopt,beopt,bdopt,messages,
         
         if reuse_m and pref+'wi'+bd_ in m_:   # input layer
             
-            m[pref+'wi'] = nn.Parameter(torch.tensor(m_[pref+'wi'+bd_]),requires_grad=True)
-            m[pref+'bi'] = nn.Parameter(torch.tensor(m_[pref+'bi'+bd_]),requires_grad=True)
-            m[pref+'wo'] = nn.Parameter(torch.tensor(m_[pref+'wo'+bd_]),requires_grad=True)
-            m[pref+'bo'] = nn.Parameter(torch.tensor(m_[pref+'bo'+bd_]),requires_grad=True)
+            m[pref+'wi'] = nn.Parameter(torch.tensor(m_[pref+'wi'+bd_],
+                                                    dtype=torch.double),requires_grad=True)
+            m[pref+'bi'] = nn.Parameter(torch.tensor(m_[pref+'bi'+bd_],
+                                                    dtype=torch.double),requires_grad=True)
+            m[pref+'wo'] = nn.Parameter(torch.tensor(m_[pref+'wo'+bd_],
+                                                    dtype=torch.double),requires_grad=True)
+            m[pref+'bo'] = nn.Parameter(torch.tensor(m_[pref+'bo'+bd_],
+                                                    dtype=torch.double),requires_grad=True)
             for i in range(layer[1]):   
-                m[pref+'w'].append(nn.Parameter(torch.tensor(m_[pref+'w'+bd_][i]),
-                                                    requires_grad=True) ) 
-                m[pref+'b'].append(nn.Parameter(torch.tensor(m_[pref+'b'+bd_][i]),
-                                                    requires_grad=True) )
+                m[pref+'w'].append(nn.Parameter(torch.tensor(m_[pref+'w'+bd_][i],
+                                                 dtype=torch.double),requires_grad=True) ) 
+                m[pref+'b'].append(nn.Parameter(torch.tensor(m_[pref+'b'+bd_][i],
+                                                 dtype=torch.double),requires_grad=True) )
         else:
             m[pref+'wi'] = nn.Parameter(torch.randn(nin,layer[0]),requires_grad=True)   
             m[pref+'bi'] = nn.Parameter(torch.randn(layer[0]),requires_grad=True)  
             m[pref+'wo'] = nn.Parameter(torch.randn(layer[0],nout),requires_grad=True)   
             m[pref+'bo'] = nn.Parameter(torch.randn(nout)+bias,requires_grad=True)
             for i in range(layer[1]):   
-                m[pref+'w'].append(nn.Parameter(torch.randn(layer[0],layer[0]),
-                                                requires_grad=True)) 
-                m[pref+'b'].append(nn.Parameter(torch.randn(layer[0]),
-                                                requires_grad=True)) 
+                m[pref+'w'].append(nn.Parameter(torch.randn(layer[0],layer[0],
+                                        dtype=torch.double),requires_grad=True)) 
+                m[pref+'b'].append(nn.Parameter(torch.randn(layer[0],
+                                        dtype=torch.double),requires_grad=True)) 
         return None  # End of local funciton definition
 
     ############ set weight and bias for message neural network ###################
